third_party/blink/web_tests/external/wpt/svg/styling/css-selectors-case-sensitivity-expected.txt

This is a testharness.js-based test.
[FAIL] Camel-cased viewBox on svg in html document matches sensitively.
  assert_equals: expected "rgba(0, 0, 0, 0)" but got "rgb(0, 128, 0)"
[FAIL] Normalized camel-cased viewBox on svg in html document matches case-sensitively.
  assert_equals: expected "rgba(0, 0, 0, 0)" but got "rgb(0, 128, 0)"
[FAIL] Camel-cased svg foreignObject in html document matches sensitively.
  assert_equals: expected "rgba(0, 0, 0, 0)" but got "rgb(0, 128, 0)"
[FAIL] Normalized camel-cased svg foreignObject in html document matches case-sensitively.
  assert_equals: expected "rgba(0, 0, 0, 0)" but got "rgb(0, 128, 0)"
Harness: the test ran to completion.
